Removed unnecessary cast

+++ crates/bevy_ui/src/widget/text.rs
@@ -378,7 +378,7 @@ pub fn text_system(
                     panic!("Fatal error when processing text: {e}.");
                 }
                 Ok(()) => {
-                    text_layout_info.scale_factor = node.inverse_scale_factor().recip() as f32;
+                    text_layout_info.scale_factor = node.inverse_scale_factor().recip();
                     text_layout_info.size *= node.inverse_scale_factor();
                     text_flags.needs_recompute = false;
                 }
